chore: move rollup plugin to utilities
