+++ utilities/rollup-plugin-mangle-internal.js
